Adding a 3DUSE-config CLI version displayer.
In order to prepare for tagged version is can be helpfull
to have at hand a command that displays (in a post install
context) the application version as well as the versions of
its dependencies upon which it was build.
The command is included by default in the tests which can
be usefull for CI debugging.
The gui application no longer polutes std::cout with an
hardcoded version number.
The application version number is set within the main
CMakeLists.txt (look for variables prefixed with PRJ_VERSION)