Change the loading process of CityGML.
In the construction process of the citygml:polygon, a check was made on each linear ring. The objective was to remove all duplicated vertices. It is relevant for the last vertex which is a repetition of the first one, but this created some incoherence with other cases, specially because of the use of a "float" epsilon with "double" coordinates.

The new version of the code do not browse the entire linear ring, it only compares the first and the last vertices and remove this is they are equal.

If they are not equal, or if there are other duplicated vertices in the ring, it means that there are an error in the CityGML file and the polygon is invalid, but this is not the purpose of our loader to check if a CityGML is well written. This loader is based on the principle that the loaded CityGML file checks the standard rules.